Simplify server setup
In order to get the SocketIO library working, I brought the routes into
the setup of the app.

This allows us to use the same server and app instances when connecting
everything.

Notice the use of `server.listen`. This allows the SocketIO library to
also serve the correct JS to the browser and allow the ws connection to
succeed.
