+++ package.json
@@ -20,6 +20,7 @@
     "body-parser": "^1.18.2",
     "express": "^4.16.2",
     "express-handlebars": "^3.0.0",
+    "nodemon": "^1.12.1",
     "redis": "^2.8.0",
     "shorthash": "0.0.2",
     "socket.io": "^2.0.3",
